backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-cors==4.0.0
flask-migrate==4.0.4
flask-socketio==5.3.6
Pillow==9.5.0
python-dotenv==1.0.0
werkzeug==2.3.6
SQLAlchemy==2.0.15
requests==2.31.0
python-magic==0.4.27
email-validator==2.0.0.post2